refactor(admin): extract FederationManagement module
Move the Federation Management section (template, state, methods, styles)
out of the 7.5k-line AdminPanel.vue into a self-contained component.
Introduce adminShared.css (common module styles) and adminFormat.ts
(shared formatters) as the base for extracting the remaining modules.
