+++ src/components/admin/adminShared.css
@@ -1,0 +1,423 @@
+/* Shared admin module styles - used by every admin/* component via
+   <style scoped src="./adminShared.css">. Source of truth extracted from
+   AdminPanel.vue during the module split. */
+
+
+
+.action-btn {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  padding: 8px 16px;
+  background: var(--background-secondary);
+  border: 1px solid var(--border-color);
+  border-radius: 8px;
+  color: var(--text-primary);
+  font-size: 14px;
+  font-weight: 500;
+  cursor: pointer;
+  transition: all 0.2s ease;
+}
+
+
+
+.action-btn:hover {
+  background: var(--background-tertiary);
+  border-color: var(--accent-color);
+  transform: translateY(-1px);
+}
+
+
+
+.action-btn:disabled {
+  opacity: 0.5;
+  cursor: not-allowed;
+  transform: none;
+}
+
+
+
+.admin-module {
+  background: var(--background-secondary);
+  border: 1px solid var(--border-color);
+  border-radius: 12px;
+  overflow: hidden;
+  min-width: 0;
+  transition: all 0.3s ease;
+}
+
+
+
+.admin-module:hover {
+  border-color: var(--accent-color);
+  box-shadow: 0 8px 32px rgba(0, 212, 255, 0.1);
+  transform: translateY(-2px);
+}
+
+
+
+.module-header {
+  display: flex;
+  align-items: center;
+  gap: 12px;
+  padding: 20px 24px;
+  background: linear-gradient(135deg, rgba(0, 212, 255, 0.05), rgba(0, 255, 136, 0.05));
+  border-bottom: 1px solid var(--border-color);
+}
+
+
+
+.module-header h2 {
+  font-size: 18px;
+  font-weight: 600;
+  margin: 0;
+  flex: 1;
+}
+
+
+
+.stats-grid {
+  display: grid;
+  grid-template-columns: repeat(2, 1fr);
+  gap: 16px;
+  padding: 24px;
+}
+
+
+
+.stat-card {
+  display: flex;
+  align-items: center;
+  gap: 16px;
+  padding: 20px;
+  background: var(--background-tertiary);
+  border: 1px solid rgba(0, 212, 255, 0.2);
+  border-radius: 8px;
+  transition: all 0.2s ease;
+}
+
+
+
+.stat-card:hover {
+  border-color: var(--accent-color);
+  box-shadow: 0 4px 16px rgba(0, 212, 255, 0.1);
+}
+
+
+
+.stat-value {
+  font-size: 24px;
+  font-weight: 700;
+  color: var(--text-primary);
+  margin-bottom: 4px;
+}
+
+
+
+.stat-label {
+  font-size: 14px;
+  color: var(--text-secondary);
+  margin-bottom: 4px;
+}
+
+
+
+.cyber-input, .cyber-textarea, .cyber-select {
+  width: 100%;
+  padding: 12px 16px;
+  background: var(--background-tertiary);
+  border: 1px solid var(--border-color);
+  border-radius: 8px;
+  color: var(--text-primary);
+  font-size: 14px;
+  transition: all 0.2s ease;
+}
+
+
+
+.cyber-input:focus, .cyber-textarea:focus, .cyber-select:focus {
+  outline: none;
+  border-color: var(--accent-color);
+  box-shadow: 0 0 0 3px rgba(0, 212, 255, 0.1);
+}
+
+
+
+/* Federation Management Styles */
+.module-actions {
+  display: flex;
+  gap: 8px;
+  margin-left: auto;
+}
+
+
+
+.primary-btn, .primary-btn-sm {
+  display: flex;
+  align-items: center;
+  gap: 6px;
+  padding: 8px 16px;
+  background: var(--accent-color);
+  border: none;
+  border-radius: 6px;
+  color: var(--text-primary);
+  font-weight: 500;
+  cursor: pointer;
+  transition: all 0.2s ease;
+}
+
+
+
+.primary-btn:hover, .primary-btn-sm:hover {
+  background: #0099cc;
+  transform: translateY(-1px);
+}
+
+
+
+.spinner-small {
+  width: 14px;
+  height: 14px;
+  border: 2px solid rgba(255, 69, 58, 0.3);
+  border-top-color: #ff453a;
+  border-radius: 50%;
+  display: inline-block;
+  animation: spin 0.8s linear infinite;
+}
+
+
+
+.maintenance-card .action-btn.primary {
+  background: var(--accent-primary);
+  border-color: var(--accent-primary);
+}
+
+
+
+.error-text {
+  color: #ff453a;
+}
+
+
+
+.loading-state {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 12px;
+  padding: 40px;
+  color: var(--text-secondary);
+}
+
+
+
+.badge {
+  padding: 2px 8px;
+  border-radius: 12px;
+  font-size: 11px;
+  font-weight: 600;
+  text-transform: uppercase;
+}
+
+
+
+.badge.trusted {
+  background: rgba(0, 255, 136, 0.2);
+  color: #00ff88;
+}
+
+
+
+.badge.blocked {
+  background: rgba(255, 69, 58, 0.2);
+  color: #ff453a;
+}
+
+
+
+.badge.inactive {
+  background: rgba(156, 163, 175, 0.2);
+  color: #9ca3af;
+}
+
+
+
+.badge.success {
+  background: rgba(0, 255, 136, 0.2);
+  color: #00ff88;
+}
+
+
+
+.badge.info {
+  background: rgba(0, 212, 255, 0.2);
+  color: #00d4ff;
+}
+
+
+
+.pagination {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 16px;
+  padding: 20px;
+  border-top: 1px solid var(--border-color);
+}
+
+
+
+.empty-state {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  gap: 16px;
+  padding: 40px;
+  text-align: center;
+  color: var(--text-secondary);
+}
+
+
+
+.search-bar .cyber-input {
+  padding-left: 36px;
+  max-width: 200px;
+}
+
+
+
+.user-name .badge {
+  font-size: 10px;
+  padding: 2px 6px;
+  border-radius: 4px;
+  font-weight: 600;
+  text-transform: uppercase;
+}
+
+
+
+.badge.suspended {
+  background: rgba(255, 193, 7, 0.2);
+  color: #ffc107;
+}
+
+
+
+.badge.admin {
+  background: rgba(0, 212, 255, 0.2);
+  color: #00d4ff;
+}
+
+
+
+.badge.moderator {
+  background: rgba(46, 204, 113, 0.2);
+  color: #2ecc71;
+}
+
+
+.announcement-item .badge.inactive { background: var(--background-quaternary); color: var(--text-muted); }
+
+@media (max-width: 768px) {
+
+  .admin-module {
+    max-width: calc(100vw - 32px);
+  }
+
+
+  /* Two-up stats read better than a single tall column on phones. */
+  .stats-grid {
+    grid-template-columns: 1fr 1fr;
+  }
+}
+
+@media (max-width: 480px) {
+
+  .admin-module {
+    max-width: calc(100vw - 24px);
+  }
+
+  .stats-grid {
+    grid-template-columns: 1fr;
+  }
+
+  .admin-actions .action-btn {
+    width: 100%;
+    justify-content: center;
+    min-height: 44px;
+  }
+}
+
+
+
+.badge.owner {
+  background: rgba(255, 193, 7, 0.2);
+  color: #ffc107;
+  font-size: 10px;
+  padding: 2px 6px;
+  border-radius: 4px;
+  font-weight: 600;
+  text-transform: uppercase;
+}
+
+
+
+.badge.sensitive {
+  background: rgba(250, 166, 26, 0.2);
+  color: #faa61a;
+}
+
+
+
+.badge.cw {
+  background: rgba(88, 101, 242, 0.2);
+  color: #7c8af5;
+}
+
+
+
+.badge.silenced {
+  background: rgba(250, 166, 26, 0.2);
+  color: #faa61a;
+}
+
+
+
+.add-tier-form .cyber-input {
+  flex: 1;
+  min-width: 100px;
+}
+
+
+
+.funding-link-row .cyber-input {
+  min-width: 0;
+}
+
+
+
+.add-supporter-form .cyber-input {
+  flex: 1;
+  min-width: 100px;
+}
+
+
+
+.supporter-search-wrapper .cyber-input {
+  width: 100%;
+}
+
+
+
+@keyframes pulse {
+  0%, 100% { opacity: 1; }
+  50% { opacity: 0.5; }
+}
+
+
+
+@keyframes spin {
+  from { transform: rotate(0deg); }
+  to { transform: rotate(360deg); }
+}
